pyproject.toml
[tool.poetry]
name = "webserver"
version = "0.1.0"
description = ""
authors = ["Fabian Siegel <fabians1999@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.61.1"
uvicorn = "^0.11.8"
tinydb = "^4.1.1"
python-dateutil = "^2.8.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.8.3"
isort = "^5.5.3"
jedi = "^0.17.2"

[tool.poetry.scripts]
poetry = 'server:run'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
